producer/requirements.txt
kafka-python==2.0.2
flask==3.0.0
werkzeug==3.0.1
python-snappy==0.6.1
prometheus-client==0.19.0